moved - page code to src/index
